requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0

scikit-learn>=1.3.0,<2.0.0

tensorflow-cpu>=2.13.0,<2.16.0

matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

Pillow>=10.0.0

streamlit>=1.25.0

PyYAML>=6.0

tqdm>=4.65.0
joblib>=1.3.0

click>=8.1.0
colorama>=0.4.6

jinja2>=3.1.0

loguru>=0.7.0

